% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parsuj_skladnie_mcda}
\alias{.parsuj_skladnie_mcda}
\title{Wewnętrzny parser składni MCDA}
\usage{
.parsuj_skladnie_mcda(skladnia)
}
\description{
Funkcja pomocnicza do interpretowania modelu MCDA
zadanego jako ciąg znaków przez użytkownika.
Zamienia zapis typu "Kryterium =~ zmienna1 + zmienna2"
na listę mapującą kryteria na zmienne składowe.
}
\keyword{internal}
