% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.rozmyj_wektor}
\alias{.rozmyj_wektor}
\title{Wewnętrzna funkcja rozmywająca (fuzzifier)}
\usage{
.rozmyj_wektor(wektor)
}
\description{
Zamienia wektor liczb rzeczywistych
na trójkątne liczby rozmyte (TFN).
Każda wartość x jest mapowana na (x-1, x, x+1).
}
\keyword{internal}
