% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_waspas_wynik}
\alias{plot.rozmyty_waspas_wynik}
\title{Mapa Spójności WASPAS}
\usage{
\method{plot}{rozmyty_waspas_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_waspas_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Porównuje podejście addytywne (WSM) z multiplikatywnym (WPM).
Jeśli punkty leżą na przekątnej, metoda jest spójna.
}
