% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_topsis}
\alias{rozmyty_topsis}
\title{Rozmyta Metoda TOPSIS}
\usage{
rozmyty_topsis(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)

rozmyty_topsis(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz ($m \times 3n$).}

\item{typy_kryteriow}{Wektor znakowy ("max" dla zysku, "min" dla kosztu).}

\item{wagi}{(Opcjonalnie) Wektor wag.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_topsis_wynik` z rankingiem.

Obiekt klasy `rozmyty_topsis_wynik` z rankingiem.
}
\description{
Implementacja Fuzzy TOPSIS. Oblicza odległość od rozwiązania idealnego
i anty-idealnego.

Implementacja Fuzzy TOPSIS. Oblicza odległość od rozwiązania idealnego
i anty-idealnego.
}
