% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_waspas}
\alias{rozmyty_waspas}
\title{Rozmyta Metoda WASPAS}
\usage{
rozmyty_waspas(
  macierz_decyzyjna,
  typy_kryteriow,
  lambda = 0.5,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)

rozmyty_waspas(
  macierz_decyzyjna,
  typy_kryteriow,
  lambda = 0.5,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz ($m \times 3n$).}

\item{typy_kryteriow}{Wektor znakowy ("max" dla zysku, "min" dla kosztu).}

\item{lambda}{Parametr wagi WSM vs WPM (domyślnie 0.5).}

\item{wagi}{(Opcjonalnie) Wektor wag.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\description{
Weighted Aggregated Sum Product Assessment. Łączy podejście addytywne (WSM)
i multiplikatywne (WPM).

Weighted Aggregated Sum Product Assessment. Łączy podejście addytywne (WSM)
i multiplikatywne (WPM).
}
