% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\docType{data}
\name{social_media_data}
\alias{social_media_data}
\title{Dane dotyczące użycia mediów społecznościowych przez studentów (MCDA)}
\format{
Ramka danych (data frame) z 200 wierszami i 7 zmiennymi:
\describe{
  \item{studentID}{Unikalny identyfikator studenta}
  \item{most_used_platform}{Główna używana platforma (Alternatywa w MCDA)}
  \item{avg_daily_usage}{Średni dzienny czas spędzony na platformie (w godzinach)}
  \item{mental_health}{Subiektywna ocena zdrowia psychicznego (skala 1-10, gdzie 10 to stan najlepszy)}
  \item{addicted_score}{Wskaźnik poziomu uzależnienia (skala 1-10)}
  \item{sleep}{Średnia liczba godzin snu na dobę}
  \item{conflicts}{Liczba konfliktów interpersonalnych wynikających z użycia social mediów (0-10)}
}
}
\source{
Zbiór danych "Students Social Media Addiction" przetworzony na potrzeby pakietu.
}
\usage{
data(social_media_data)
}
\description{
Zbiór danych zawierający rekordy dotyczące wpływu najpopularniejszych platform
społecznościowych na życie studentów (Graduate i Undergraduate).
Dane obejmują platformy: Facebook, TikTok, Instagram oraz WhatsApp.
Przeznaczony do testowania metod MCDA (Multi-Criteria Decision Analysis).
}
\details{
Dane pochodzą ze zbioru "Students Social Media Addiction". Zostały one przefiltrowane,
aby uwzględnić wyłącznie osoby na poziomie akademickim "Graduate" i "Undergraduate"
oraz najpopularniejsze platformy komunikacyjne.

Obiekt został przygotowany w procesie inżynierii danych tak, aby nazwy kolumn były
kompatybilne z funkcjami pakietu takimi jak \code{przygotuj_dane_mcda}.
}
\keyword{datasets}
