% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie tabeli w standardzie APA}
\usage{
tabela_apa(meta_wynik)
}
\arguments{
\item{meta_wynik}{Obiekt listy zwracany przez funkcję `rozmyty_meta_ranking`.}
}
\value{
Obiekt klasy `flextable` sformatowany zgodnie z zasadami APA.
}
\description{
Konwertuje wyniki zbiorcze z procedury rozmytego meta-rankingu
do sformatowanej tabeli zgodnej z wymogami manuala APA (7th edition), wzorem z pracy
oraz czcionką Times New Roman.
}
